Heck rust treating fn items and fn pointers as distict types.
Also heck writing memcpy (et. al) in assembly

+++ kernel/common/src/errno.rs
@@ -1,3 +1,3 @@
 
 
-pub const ENOSYS: i32 = 1024; // This is hardcoded by arch/x86_64/syscall_init.s, which makes syscalls
+pub const ENOSYS: isize = 1024; // This is hardcoded by arch/x86_64/syscall_init.s, which makes syscalls
